pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localflow"
version = "0.1.0"
description = "Local speech-to-text with GPU transcription and LLM refinement"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6",
    "sounddevice>=0.4",
    "faster-whisper>=1.0",
    "httpx>=0.27",
    "numpy>=1.26",
    "python-dotenv>=1.0",
]

[project.scripts]
localflow = "localflow.__main__:main"

[tool.setuptools.packages.find]
include = ["localflow*"]
